+++ src/builtins/mod.rs
@@ -1,5 +1,7 @@
+mod crouch;
 mod jump;
 mod walk;
 
+pub use crouch::{TnuaBuiltinCrouch, TnuaBuiltinCrouchState};
 pub use jump::{TnuaBuiltinJump, TnuaBuiltinJumpState};
 pub use walk::{TnuaBuiltinWalk, TnuaBuiltinWalkState};