+++ src/builtins/walk.rs
@@ -2,6 +2,7 @@ use std::time::Duration;
 
 use bevy::prelude::*;
 
+use crate::basis_action_traits::TnuaBasisContext;
 use crate::util::ProjectionPlaneForRotation;
 use crate::{TnuaBasis, TnuaVelChange};
 
@@ -13,8 +14,6 @@ pub struct TnuaBuiltinWalk {
     pub up: Vec3,
     pub spring_strengh: f32,
     pub spring_dampening: f32,
-    pub height_change_impulse_for_duration: f32,
-    pub height_change_impulse_limit: f32,
     pub acceleration: f32,
     pub air_acceleration: f32,
     pub coyote_time: f32,
@@ -28,23 +27,13 @@ impl TnuaBasis for TnuaBuiltinWalk {
     const NAME: &'static str = "TnuaBuiltinWalk";
     type State = TnuaBuiltinWalkState;
 
-    fn apply(
-        &self,
-        state: &mut Self::State,
-        ctx: crate::basis_action_traits::TnuaBasisContext,
-        motor: &mut crate::TnuaMotor,
-    ) {
+    fn apply(&self, state: &mut Self::State, ctx: TnuaBasisContext, motor: &mut crate::TnuaMotor) {
         match &mut state.airborne_state {
             AirborneState::NoJump => {}
             AirborneState::FreeFall { coyote_time } => {
                 coyote_time.tick(Duration::from_secs_f32(ctx.frame_duration));
             }
         }
-        // TODO: calc the climb vectors
-
-        let float_height_offset = 0.0; // is this needed?
-        let prev_float_height_offset = 0.0; // maybe this should be prev_float_height instead of an
-                                            // offset?
 
         let climb_vectors: Option<ClimbVectors>;
         let considered_in_air: bool;
@@ -135,43 +124,8 @@ impl TnuaBasis for TnuaBuiltinWalk {
                             // not doing the jump calculation here
                             let spring_offset = self.float_height - sensor_output.proximity;
                             state.standing_offset = -spring_offset;
-                            let spring_offset = spring_offset + float_height_offset;
-                            let spring_force: f32 = spring_offset * self.spring_strengh;
-                            let offset_change_impulse: f32 =
-                                if 0.01 <= (float_height_offset - prev_float_height_offset).abs() {
-                                    let velocity_to_get_to_new_float_height =
-                                        spring_offset / self.height_change_impulse_for_duration;
-                                    velocity_to_get_to_new_float_height.clamp(
-                                        -self.height_change_impulse_limit,
-                                        self.height_change_impulse_limit,
-                                    )
-                                } else {
-                                    0.0
-                                };
-
-                            let relative_velocity =
-                                state.effective_velocity.dot(self.up) - state.vertical_velocity;
-
-                            let dampening_force =
-                                relative_velocity * self.spring_dampening / ctx.frame_duration;
-                            let spring_force = spring_force - dampening_force;
-
-                            let gravity_compensation = -ctx.tracker.gravity.dot(self.up);
-
-                            let spring_impulse =
-                                ctx.frame_duration * (spring_force + gravity_compensation);
-
-                            let impulse_to_use =
-                                if spring_impulse.abs() < offset_change_impulse.abs() {
-                                    offset_change_impulse
-                                } else {
-                                    spring_impulse
-                                };
-
-                            // TODO: maybe this needs to be an acceleration rather than an
-                            // impulse? The problem is the comparison between `spring_impulse`
-                            // and `offset_change_impulse`...
-                            break 'upward_impulse TnuaVelChange::boost(impulse_to_use * self.up);
+                            let boost = self.spring_force_boost(state, &ctx, spring_offset);
+                            break 'upward_impulse TnuaVelChange::boost(boost * self.up);
                         } else {
                             state.airborne_state = AirborneState::FreeFall {
                                 coyote_time: Timer::new(
@@ -246,8 +200,6 @@ impl TnuaBasis for TnuaBuiltinWalk {
     }
 
     fn proximity_sensor_cast_range(&self) -> f32 {
-        // TODO - also need to consider float_height_offset? Or maybe it should be united,
-        // or converted into an action?
         self.float_height + self.cling_distance
     }
 
@@ -268,6 +220,29 @@ impl TnuaBasis for TnuaBuiltinWalk {
 
     fn vertical_velocity(&self, state: &Self::State) -> f32 {
         state.vertical_velocity
+    }
+}
+
+impl TnuaBuiltinWalk {
+    // TODO: maybe this needs to be an acceleration rather than an
+    // impulse? The problem is the comparison between `spring_impulse`
+    // and `offset_change_impulse`...
+    pub fn spring_force_boost(
+        &self,
+        state: &TnuaBuiltinWalkState,
+        ctx: &TnuaBasisContext,
+        spring_offset: f32,
+    ) -> f32 {
+        let spring_force: f32 = spring_offset * self.spring_strengh;
+
+        let relative_velocity = state.effective_velocity.dot(self.up) - state.vertical_velocity;
+
+        let dampening_force = relative_velocity * self.spring_dampening / ctx.frame_duration;
+        let spring_force = spring_force - dampening_force;
+
+        let gravity_compensation = -ctx.tracker.gravity.dot(self.up);
+
+        ctx.frame_duration * (spring_force + gravity_compensation)
     }
 }